test(arrival): B2 I1 — test-tsconfig honest-type fixes, 43 → 20 errors
23 mechanical fixes across 15 test files: stale casts deleted, Dirent
retype, 3 new graph-layer LineageNode kinds in the checkpoint switch,
stale @ts-expect-error removed, readonly-array widenings, resolveSymbols
retyped to the authoring shape, public ZodTuple narrowing instead of
.def reach-through, toSorted over readonly, scheme-space void vs
JS-space [] on contract-typed impls. Remaining 20 = the RULED exec
cluster (16, lands with the core tranche) + 4 flagged source-territory
items (ADict entry type too narrow for settleEntry's documented
behavior; AUnwrap missing ABytevector arm; AListAlike I1c judgment).
Build gate 0 errors throughout; full suite green minus the known
wall-clock flake.

+++ packages/arrival/src/__tests__/scheme-vector-serialization.test.ts
@@ -33,7 +33,11 @@ describe("boxed vector/bytevector — Scheme→JS serialization (schemeToJs)", (
     const bv = new ABytevector(CONSTANT_CTX, Uint8Array.from([4, 5, 6]));
     const out = schemeToJs(bv);
     expect(out).toBeInstanceOf(Uint8Array);
-    expect([...out]).toEqual([4, 5, 6]);
+    // Cast: `AUnwrap<T>` (values/types.ts) has no `ABytevector` arm — falls through to
+    // `unknown` — even though rosetta.ts's schemeToJsImpl documents "ABytevector → raw
+    // Uint8Array" (a source-type gap, not fixed here per I1 scope: flagged separately).
+    // The `toBeInstanceOf` assertion above is the runtime proof this narrows honestly.
+    expect([...(out as Uint8Array)]).toEqual([4, 5, 6]);
   });
 });
